move elastic testcase to child conftest
